requirements.txt
opencv-python>=4.8.0
mediapipe==0.10.14
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0,<2
